Apply singleton pattern to Config object
The config object is currently a global within khard.py. Applying the
singleton pattern is a more reliable and explicit method of passing the
configuration information.

The first time the Config object is instantiated it will load the
configuration file and place the object in memory. Subsequent functions
that require the config information will also instantiate the object,
but the object will recognize that it already exists in memory and
return the existing object rather than creating a new separate instance.