Separate the concepts of print and export in CarddavObject
In prior code the print_vcard method in CarddavObject was used for both
displaying to the screen and text-based search. This prevents the use of
custom display templates. This commit therefore separates those
functions into two separate methods:

* print_vcard() method is responsible for printing to the screen

* export_vcard() method returns a string in the "standard" format for
  text-based search.

The "standard" format is presently the same format as previously output
by the print_vcard method.