Add rolodex print format
Pretty prints the details of a contact in the form of a rolodex card.
Automatically adjusts for narrow terminal widths. The format can be
changed via the print_template setting in the config file. Also added a
post_format option to format post addresses in US format. EU format is
the original address formatting.